Plan: plans/PR-Reasoning-Enrichment-Pack-Split.md
Split atlas review enrichment into product pack

Intentional:
- Preserve atlas_brain.reasoning.evidence_engine.EvidenceEngine as the existing public object shape.
- Keep extracted_reasoning_core evidence engine slim.

Deferred:
- Exporting the enrichment pack for other products.
- Graph/state slimming.
